+++ SampleGame/STR/Perlin.cpp
@@ -76,8 +76,8 @@ namespace Perlin
 
 	float Perlin(float x, float y)
 	{
-		x = std::fmodf(x, 255.f);
-		y = std::fmodf(y, 255.f);
+		x = std::fmod(x, 255.f);
+		y = std::fmod(y, 255.f);
 
 		int xi = (int)x & 255;
 		int yi = (int)y & 255;
@@ -109,9 +109,9 @@ namespace Perlin
 
 	float Perlin3(float x, float y, float z)
 	{
-		x = std::fmodf(x, 255.f);
-		y = std::fmodf(y, 255.f);
-		z = std::fmodf(z, 255.f);
+		x = std::fmod(x, 255.f);
+		y = std::fmod(y, 255.f);
+		z = std::fmod(z, 255.f);
 
 		int xi = (int)x & 255;
 		int yi = (int)y & 255;
